Move PrefixWrappedAnalyzer into private class
